version 4.1.0

Conflicts:
	Gemfile.lock
	app/controllers/omniauth_callbacks_controller.rb
	app/views/profiles/account.html.haml
	app/views/projects/_clone_panel.html.haml
	config/gitlab.yml.example
	config/initializers/1_settings.rb
	config/initializers/devise.rb
	lib/gitlab/auth.rb

+++ app/controllers/projects_controller.rb
@@ -1,7 +1,8 @@
-require Rails.root.join('lib', 'gitlab', 'graph_commit')
+require Rails.root.join('lib', 'gitlab', 'graph', 'json_builder')
 
 class ProjectsController < ProjectResourceController
   skip_before_filter :project, only: [:new, :create]
+  skip_before_filter :repository, only: [:new, :create]
 
   # Authorize
   before_filter :authorize_read_project!, except: [:index, :new, :create]
@@ -18,12 +19,13 @@ class ProjectsController < ProjectResourceController
   end
 
   def create
-    @project = Project.create_by_user(params[:project], current_user)
+    @project = Projects::CreateContext.new(current_user, params[:project]).execute
 
     respond_to do |format|
+      flash[:notice] = 'Project was successfully created.' if @project.saved?
       format.html do
         if @project.saved?
-          redirect_to(@project, notice: 'Project was successfully created.')
+          redirect_to @project
         else
           render action: "new"
         end
@@ -33,8 +35,11 @@ class ProjectsController < ProjectResourceController
   end
 
   def update
+    status = Projects::UpdateContext.new(project, current_user, params).execute
+
     respond_to do |format|
-      if project.update_attributes(params[:project])
+      if status
+        flash[:notice] = 'Project was successfully updated.'
         format.html { redirect_to edit_project_path(project), notice: 'Project was successfully updated.' }
         format.js
       else
@@ -42,6 +47,10 @@ class ProjectsController < ProjectResourceController
         format.js
       end
     end
+
+  rescue Project::TransferError => ex
+    @error = ex
+    render :update_failed
   end
 
   def show
@@ -50,12 +59,12 @@ class ProjectsController < ProjectResourceController
 
     respond_to do |format|
       format.html do
-         unless @project.empty_repo?
-           @last_push = current_user.recent_push(@project.id)
-           render :show
-         else
-           render "projects/empty"
-         end
+        if @project.repository && !@project.repository.empty?
+          @last_push = current_user.recent_push(@project.id)
+          render :show
+        else
+          render "projects/empty"
+        end
       end
       format.js
     end
@@ -71,7 +80,10 @@ class ProjectsController < ProjectResourceController
 
   def wall
     return render_404 unless @project.wall_enabled
-    @note = Note.new
+
+    @target_type = :wall
+    @target_id = nil
+    @note = @project.notes.new
 
     respond_to do |format|
       format.html
@@ -79,15 +91,22 @@ class ProjectsController < ProjectResourceController
   end
 
   def graph
-    @days_json, @commits_json = Gitlab::GraphCommit.to_graph(project)
+    respond_to do |format|
+      format.html
+      format.json do
+        graph = Gitlab::Graph::JsonBuilder.new(project)
+        render :json => graph.to_json
+      end
+    end
   end
 
   def destroy
-    # Disable the UsersProject update_repository call, otherwise it will be
-    # called once for every person removed from the project
-    UsersProject.skip_callback(:destroy, :after, :update_repository)
+    return access_denied! unless can?(current_user, :remove_project, project)
+
+    # Delete team first in order to prevent multiple gitolite calls
+    project.team.truncate
+
     project.destroy
-    UsersProject.set_callback(:destroy, :after, :update_repository)
 
     respond_to do |format|
       format.html { redirect_to root_path }
